feat: update deployment for upgrades
