+++ packages/contracts/deploy/001-OVM_ChainStorageContainer_ctc_batches.deploy.ts
@@ -22,6 +22,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['ChainStorageContainer_ctc_batches']
+deployFn.tags = ['ChainStorageContainer_ctc_batches', 'upgrade']
 
 export default deployFn
