+++ packages/contracts/deploy/002-OVM_ChainStorageContainer_ctc_queue.deploy.ts
@@ -22,6 +22,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['ChainStorageContainer_ctc_queue']
+deployFn.tags = ['ChainStorageContainer_ctc_queue', 'upgrade']
 
 export default deployFn
