+++ packages/contracts/deploy/003-OVM_ChainStorageContainer_scc_batches.deploy.ts
@@ -22,6 +22,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['ChainStorageContainer_scc_batches']
+deployFn.tags = ['ChainStorageContainer_scc_batches', 'upgrade']
 
 export default deployFn
