+++ packages/contracts/deploy/004-OVM_CanonicalTransactionChain.deploy.ts
@@ -26,6 +26,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['CanonicalTransactionChain']
+deployFn.tags = ['CanonicalTransactionChain', 'upgrade']
 
 export default deployFn
