+++ packages/contracts/deploy/005-OVM_StateCommitmentChain.deploy.ts
@@ -25,6 +25,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['StateCommitmentChain']
+deployFn.tags = ['StateCommitmentChain', 'upgrade']
 
 export default deployFn
