+++ packages/contracts/deploy/006-OVM_BondManager.deploy.ts
@@ -21,6 +21,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['BondManager']
+deployFn.tags = ['BondManager', 'upgrade']
 
 export default deployFn
