+++ packages/contracts/deploy/008-Proxy__OVM_L1CrossDomainMessenger.deploy.ts
@@ -20,7 +20,7 @@ const deployFn: DeployFunction = async (hre) => {
     name: 'Proxy__L1CrossDomainMessenger',
     contract: 'Lib_ResolvedDelegateProxy',
     iface: 'L1CrossDomainMessenger',
-    args: [Lib_AddressManager.address, 'L1CrossDomainMessenger'],
+    args: [Lib_AddressManager.address, 'OVM_L1CrossDomainMessenger'],
     postDeployAction: async (contract) => {
       console.log(`Initializing Proxy__L1CrossDomainMessenger...`)
       await contract.initialize(Lib_AddressManager.address)
