+++ packages/contracts/deploy/012-finalize.ts
@@ -39,6 +39,6 @@ const deployFn: DeployFunction = async (hre) => {
 }
 
 deployFn.dependencies = ['Lib_AddressManager']
-deployFn.tags = ['finalize']
+deployFn.tags = ['finalize', 'upgrade']
 
 export default deployFn
